enable parallel execution of integration tests
